third_party/WebKit/LayoutTests/fast/scroll-behavior/scroll-customization/scrollstate-consume-deltas.html

<!DOCTYPE html>
<html>
<head>
<title>ScrollState consumeDelta</title>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
</head>
<body>
<script>

function deltaShouldBe(scrollState, x, y, testName) {
  test(function() {
    assert_equals(scrollState.deltaX, x);
    assert_equals(scrollState.deltaY, y);
  }, testName);
}

test(function() {
  assert_true('ScrollState' in window, "'ScrollState' in window");
}, "These tests only work with scroll customization enabled.");

if ('ScrollState' in window) {
  var scrollState = new ScrollState({deltaX: 10, deltaY: -20});
  deltaShouldBe(scrollState, 10, -20, "Unconsumed deltas");
  scrollState.consumeDelta(2, -3);
  deltaShouldBe(scrollState, 8, -17, "Partially consumed deltas");
  scrollState.consumeDelta(8, -17);
  deltaShouldBe(scrollState, 0, 0, "Fully consumed deltas");
}
</script>
</body>
</html>
